# Optimize first/last when no group by interval is present