The `first()` and `last()` functions response rate would increase linear
to the number of points even though it seems like it shouldn't. This
optimization greatly reduces the amount of time to return a response
when no `GROUP BY time(...)` clause is present in a query.

```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkEngine_CreateIterator_First_1K-8       31738         28327         -10.75%
BenchmarkEngine_CreateIterator_First_100K-8     861782        72030         -91.64%
BenchmarkEngine_CreateIterator_First_1M-8       8599111       88630         -98.97%
BenchmarkEngine_CreateIterator_Last_1K-8        25177         21775         -13.51%
BenchmarkEngine_CreateIterator_Last_100K-8      828121        62467         -92.46%
BenchmarkEngine_CreateIterator_Last_1M-8        8311886       76800         -99.08%

benchmark                                       old allocs     new allocs     delta
BenchmarkEngine_CreateIterator_First_1K-8       52             62             +19.23%
BenchmarkEngine_CreateIterator_First_100K-8     68             78             +14.71%
BenchmarkEngine_CreateIterator_First_1M-8       161            171            +6.21%
BenchmarkEngine_CreateIterator_Last_1K-8        52             62             +19.23%
BenchmarkEngine_CreateIterator_Last_100K-8      68             78             +14.71%
BenchmarkEngine_CreateIterator_Last_1M-8        161            171            +6.21%

benchmark                                       old bytes     new bytes     delta
BenchmarkEngine_CreateIterator_First_1K-8       11285         12133         +7.51%
BenchmarkEngine_CreateIterator_First_100K-8     54920         55765         +1.54%
BenchmarkEngine_CreateIterator_First_1M-8       69516         70357         +1.21%
BenchmarkEngine_CreateIterator_Last_1K-8        11285         12133         +7.51%
BenchmarkEngine_CreateIterator_Last_100K-8      54919         55765         +1.54%
BenchmarkEngine_CreateIterator_Last_1M-8        69536         70357         +1.18%
```

Fixes #6704.