@jwilder 2016-10-21
Do you have any benchmarks you could add to the desc?