@jsternberg 2016-10-21
Updated the description above with benchmarks.