@jsternberg 2016-10-21
I changed the benchmark to use `GROUP BY host` since I thought it was a better test. Here we go:

```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkEngine_CreateIterator_First_1K-8       38842         26327         -32.22%
BenchmarkEngine_CreateIterator_First_100K-8     1669364       76961         -95.39%
BenchmarkEngine_CreateIterator_First_1M-8       17100447      79329         -99.54%
BenchmarkEngine_CreateIterator_Last_1K-8        33950         21437         -36.86%
BenchmarkEngine_CreateIterator_Last_100K-8      1631698       59007         -96.38%
BenchmarkEngine_CreateIterator_Last_1M-8        17099451      79239         -99.54%

benchmark                                       old allocs     new allocs     delta
BenchmarkEngine_CreateIterator_First_1K-8       158            70             -55.70%
BenchmarkEngine_CreateIterator_First_100K-8     10074          86             -99.15%
BenchmarkEngine_CreateIterator_First_1M-8       100167         179            -99.82%
BenchmarkEngine_CreateIterator_Last_1K-8        158            70             -55.70%
BenchmarkEngine_CreateIterator_Last_100K-8      10074          86             -99.15%
BenchmarkEngine_CreateIterator_Last_1M-8        100167         179            -99.82%

benchmark                                       old bytes     new bytes     delta
BenchmarkEngine_CreateIterator_First_1K-8       13232         12512         -5.44%
BenchmarkEngine_CreateIterator_First_100K-8     215267        56144         -73.92%
BenchmarkEngine_CreateIterator_First_1M-8       1669911       70736         -95.76%
BenchmarkEngine_CreateIterator_Last_1K-8        13232         12512         -5.44%
BenchmarkEngine_CreateIterator_Last_100K-8      215269        56144         -73.92%
BenchmarkEngine_CreateIterator_Last_1M-8        1669896       70736         -95.76%
```

Leaving the previous benchmark up anyway.